config/requirements.txt
# Core dependencies
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
idna==3.10
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.8.5
PyYAML>=5.4.1
discord.py>=2.3.2
discord>=2.3.2
mwparserfromhell>=0.6.5

# Discord integration
discord-py-interactions>=4.4.0
discord-py-slash-command>=4.2.1
discord-webhook>=1.1.0
websockets>=10.4
PyNaCl>=1.5.0
pillow>=9.5.0
discord.py-self>=1.9.2
discord-ext-ipc>=2.1.1

# Pokemon data sources
pokebase>=1.3.0
pokepy>=0.6.2
pokemontcgsdk>=2.0.0
beautifulsoup4>=4.12.2

# Plex integration
plexapi>=4.15.4
tmdbsimple>=2.9.1
IMDbPY>=2021.4.18

# GitHub integration
PyGithub==2.1.1
stackapi==0.3.1
vdf==3.4

# Spotify integration
spotipy==2.23.0

# YouTube integration
google-api-python-client==2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Data processing
pandas>=1.3.5
beautifulsoup4>=4.12.2

# Utilities
python-dotenv>=1.0.0
rich==13.6.0
prometheus_client>=0.16.0
asyncio>=3.4.3

# Database
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
asyncpg>=0.25.0
aiosqlite>=0.19.0

# API and Integration
mwclient>=0.10.1
python-dateutil>=2.8.2
jsonschema>=4.20.0
aiofiles>=0.8.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
mypy==1.6.1
flake8==6.1.0
pre-commit>=2.17.0
astroid>=2.9.0
pylint>=2.12.0
zipp>=3.6.0

# Documentation
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0

# Type stubs
types-python-dotenv>=0.1.0

# Utility packages
pytz>=2023.3.post1

# New dependencies
redis==5.0.1
python-gitlab==4.1.1
GitPython==3.1.37

# Game data parsing
construct==2.10.68
bitstring==4.1.2

# API integrations
tqdm==4.66.1
colorama>=0.4.6

# OSRS Wiki parsing
mwparserfromhell>=0.6.5  # For wiki text parsing
beautifulsoup4>=4.12.2   # For web scraping
lxml>=4.9.3             # For faster BS4 parsing
aiohttp>=3.9.1          # For async HTTP requests
aiofiles>=0.8.0         # For async file operations

# External APIs
python-twitch-client>=0.7.1    # For Twitch API
google-api-python-client>=2.0.0 # For YouTube API
google-auth>=2.0.0             # For YouTube API auth
google-auth-oauthlib>=0.4.1    # For YouTube API auth
steam>=1.4.4                   # For Steam API

# New dependencies
openai>=1.3.5
typing-extensions>=4.8.0
pyyaml>=6.0.1

# Web Server
Flask>=2.0.0
Flask-CORS>=3.0.0
Flask-SocketIO>=5.0.0
gunicorn>=20.0.0
eventlet>=0.30.0

# Authentication & Caching
PyJWT>=2.0.0
bcrypt>=3.2.0

# Utilities
requests>=2.26.0
aiohttp>=3.8.0
websockets>=10.0
python-socketio>=5.0.0

# Testing
pytest>=6.0.0
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0
